+++ lib/event.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const EE = require('events');
+const events = new EE();
+
+
+module.exports = events;
