+++ package.json
@@ -4,7 +4,10 @@
   "description": "",
   "main": "chatroom.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js",
+    "lint": "eslint **/*.js",
+    "test": "jest --verbose --coverage",
+    "test-watch": "jest --watchAll --verbose --coverage"
   },
   "keywords": [],
   "author": "",
